improve types and move them around a bit
