+++ jsapp/js/account/account.constants.ts
@@ -1,19 +1,27 @@
-export type UserFieldName =
-  | 'name'
-  | 'organization'
-  | 'organization_website'
-  | 'sector'
-  | 'gender'
-  | 'bio'
-  | 'city'
-  | 'country'
-  | 'require_auth'
-  | 'twitter'
-  | 'linkedin'
-  | 'instagram';
+/**
+ * This is a list of user metadata fields known by Front end code. If you happen
+ * to add new fields, please updated this interface first :)
+ */
+export interface AccountFieldsValues {
+  name: string;
+  organization: string;
+  organization_website: string;
+  sector: string;
+  gender: string;
+  bio: string;
+  city: string;
+  country: string;
+  require_auth: boolean;
+  twitter: string;
+  linkedin: string;
+  instagram: string;
+}
 
-type UserFieldNames = {[P in UserFieldName]: UserFieldName};
-export const USER_FIELD_NAMES: UserFieldNames = {
+export type AccountFieldsErrors = {[name in UserFieldName]?: string};
+
+export type UserFieldName = keyof AccountFieldsValues;
+
+export const USER_FIELD_NAMES: Record<UserFieldName, UserFieldName> = {
   name: 'name',
   organization: 'organization',
   organization_website: 'organization_website',
